normalizer/requirements.txt
python-dotenv
pyzmq
pymongo
pyyaml